test: repository test code 작성

+++ backend/src/main/java/com/e103/ohmyguide/domain/attractionvector/entity/AttractionVector.java
@@ -16,7 +16,7 @@ public class AttractionVector extends BaseEntity {
 
     @Id
     @Column(name = "attr_id")
-    private Integer attrId;
+    private Long attrId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @MapsId
